requirements.txt
click==8.0.1
colorama==0.4.4
cycler==0.10.0
Flask==2.0.1
Flask-BasicAuth==0.2.0
importlib-metadata==4.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
mysql-connector-python==8.0.26
numpy==1.21.1
pandas==1.3.1
Pillow==8.3.1
protobuf==3.17.3
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.17.1
pytz==2021.1
six==1.16.0
typing-extensions==3.10.0.0
Werkzeug==2.0.1
wordcloud==1.8.1
zipp==3.4.1
